+++ src/Model/Request/Wallet/WalletChangeTransport.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace KassaCom\SDK\Model\Request\Wallet;
+
+use KassaCom\SDK\Model\Request\AbstractRequestTransport;
+
+class WalletChangeTransport extends AbstractRequestTransport
+{
+    public function getPath()
+    {
+        return 'wallet/change';
+    }
+}